\name{MKdescr-package}
\alias{MKdescr-package}
\alias{MKdescr}
\docType{package}
\title{
Descriptive Statistics.
}
\description{
Contains several functions for descriptive statistics that are not available 
in standard R packages such as standardized IQR, Huber-type skipped mean, 
robust coefficient of variation (CV), robust signal to noise ratio (SNR), 
as well as functions that support graphical visualization such as boxplots 
based on quartiles (not hinges), negative logarithms and generalized 
logarithms for ggplots.
}
\details{
\tabular{ll}{
Package: \tab MKdescr\cr
Type: \tab Package\cr
Version: \tab 0.4\cr
Date: \tab 2019-11-06\cr
Depends: \tab R(>= 3.5.0)\cr
Imports: \tab stats, graphics, ggplot2, scales\cr
Suggests: \tab knitr, rmarkdown\cr
License: \tab LGPL-3\cr
URL: \tab http://www.stamats.de/\cr
}
library(MKmisc)
}
\author{ Matthias Kohl \url{http://www.stamats.de}

Maintainer: Matthias Kohl  \email{matthias.kohl@stamats.de}}
%\references{
%~~ Literature or other references for background information ~~
%}
%~~ Optionally other keywords from doc/KEYWORDS, one per line
%\seealso{
%~~ Optional links to other man pages, e.g. ~~
%~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
%}
%\examples{
%~~ simple examples of the most important functions ~~
%}
\keyword{package}
