\name{MKdescr-package}
\alias{MKdescr-package}
\alias{MKdescr}
\docType{package}
\title{
Descriptive Statistics.
}
\description{
Computation of standardized interquartile range (IQR), Huber-type skipped mean 
(Hampel (1985), <doi:10.2307/1268758>), robust coefficient of variation (CV) 
(Arachchige et al. (2019), <doi:10.48550/arXiv.1907.01110>), robust signal to noise ratio (SNR), 
z-score, standardized mean difference (SMD), as well as functions that support 
graphical visualization such as boxplots based on quartiles (not hinges), negative 
logarithms and generalized logarithms for 'ggplot2' (Wickham (2016), ISBN:978-3-319-24277-4).
}
\details{
library(MKdescr)
}
\author{ Matthias Kohl \url{https://www.stamats.de}

Maintainer: Matthias Kohl  \email{matthias.kohl@stamats.de}}
%\references{
%~~ Literature or other references for background information ~~
%}
%~~ Optionally other keywords from doc/KEYWORDS, one per line
%\seealso{
%~~ Optional links to other man pages, e.g. ~~
%~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
%}
%\examples{
%~~ simple examples of the most important functions ~~
%}
\keyword{package}
