\name{MKdescr-package}
\alias{MKdescr-package}
\alias{MKdescr}
\docType{package}
\title{
Descriptive Statistics.
}
\description{
Computation of standardized interquartile range (IQR), Huber-type skipped mean 
(Hampel (1985), <doi:10.2307/1268758>), robust coefficient of variation (CV) 
(Arachchige et al. (2019), <arXiv:1907.01110>), robust signal to noise ratio (SNR), 
z-score, standardized mean difference (SMD), as well as functions that support 
graphical visualization such as boxplots based on quartiles (not hinges), negative 
logarithms and generalized logarithms for 'ggplot2' (Wickham (2016), ISBN:978-3-319-24277-4).
}
\details{
\tabular{ll}{
Package: \tab MKdescr\cr
Type: \tab Package\cr
Version: \tab 0.6\cr
Date: \tab 2020-10-18\cr
Depends: \tab R(>= 3.5.0)\cr
Imports: \tab stats, graphics, ggplot2, scales\cr
Suggests: \tab knitr, rmarkdown\cr
License: \tab LGPL-3\cr
URL: \tab http://www.stamats.de/\cr
}
library(MKmisc)
}
\author{ Matthias Kohl \url{http://www.stamats.de}

Maintainer: Matthias Kohl  \email{matthias.kohl@stamats.de}}
%\references{
%~~ Literature or other references for background information ~~
%}
%~~ Optionally other keywords from doc/KEYWORDS, one per line
%\seealso{
%~~ Optional links to other man pages, e.g. ~~
%~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
%}
%\examples{
%~~ simple examples of the most important functions ~~
%}
\keyword{package}
